# openapi.yaml
openapi: 3.1.0
info:
  title: Google Drive, Slack Integration API and ElevenLabs
  version: 1.0.0
  description: API to integrate Google Drive functionality, Slack messaging and ElevanLabs to generate Speech based on the Text
servers:
  - url: https://f859-2407-d000-1a-4cf4-f9cf-5b9d-8f18-22f0.ngrok-free.app
    description: Development server
paths:
  /privacy:
    get:
      summary: Returns the privacy policy
      operationId: getPrivacyPolicy
      responses:
        '200':
          description: Privacy policy returned successfully.
          content:
            text/html:
              schema:
                type: string

  /list_channels:
    get:
      summary: List all Slack channels
      operationId: listSlackChannels
      responses:
        '200':
          description: List of channels retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  channels:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
        '500':
          description: Error retrieving channels.

  /startAuth:
    get:
      summary: Start Google Drive authentication
      operationId: startGoogleDriveAuth
      responses:
        '200':
          description: Authentication URL provided successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  auth_url:
                    type: string

  /handleAuth:
    get:
      summary: Handle Google Drive authentication response
      operationId: handleGoogleDriveAuth
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /listFiles:
    get:
      summary: List Google Docs-compatible files
      operationId: listGoogleDriveFiles
      responses:
        '200':
          description: List of Google Docs-compatible files retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        mimeType:
                          type: string
        '500':
          description: Error retrieving files.

  /readDoc:
    get:
      summary: Read content from a Google Docs file
      operationId: readGoogleDoc
      parameters:
        - name: document_id
          in: query
          required: true
          schema:
            type: string
          description: ID of the Google Docs file.
      responses:
        '200':
          description: File content retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
        '400':
          description: Document ID is missing or invalid.
        '500':
          description: Error reading the document.

  /updateDoc:
    post:
      summary: Update a Google Docs file
      operationId: updateGoogleDoc
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                document_id:
                  type: string
                content:
                  type: string
                location_index:
                  type: integer
                  default: 1
      responses:
        '200':
          description: Document updated successfully.
        '400':
          description: Missing or invalid input data.
        '500':
          description: Error updating the document.

  /shareFileOnSlack:
    post:
      summary: Share a Google Drive document on Slack
      operationId: shareDriveFileOnSlack
      requestBody:
        description: Information required to share a Google Drive document on Slack
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channel_id:
                  type: string
                  description: The ID of the Slack channel to share the file link in
                  example: C12345678
                document_id:
                  type: string
                  description: The ID of the Google Drive document
                  example: 1aBcDeFgHiJkLmNoPqRsTuVwXyZ
                comment:
                  type: string
                  description: Optional comment to include with the file link
                  example: "Check out this document"
      responses:
        '200':
          description: Google Drive file link shared successfully on Slack
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message_id:
                    type: string
                    description: The Slack message timestamp
                    example: "1627382910.000200"
                  public_url:
                    type: string
                    description: The public URL of the Google Drive document
                    example: https://drive.google.com/file/d/1aBcDeFgHiJkLmNoPqRsTuVwXyZ/view
                  message:
                    type: string
                    description: A success message
                    example: "Google Drive file link shared successfully"
        '400':
          description: Missing or invalid input data
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "channel_id and document_id are required"
        '500':
          description: Error encountered while processing the request
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: false
                  error:
                    type: string
                    description: Error message
                    example: "Slack error: some error message"
  /generate-audio:
    post:
      operationId: generateAudioFromText
      summary: Generate Audio from Text
      description: Accepts a text prompt, sends it to ElevenLabs API, and returns the generated audio file URL.
      requestBody:
        required: true
        description: JSON payload containing the text to be converted into speech.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAudioRequest'
      responses:
        '200':
          description: Audio file generated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Audio file generated successfully."
                  file_url:
                    type: string
                    format: uri
                    example: "http://127.0.0.1:5000/audio/output.mp3"
        '400':
          description: Bad request. Text input is missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error or API failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /audio/{filename}:
    get:
      operationId: serveAudioFile
      summary: Retrieve an audio file
      description: Serves an audio file by its filename from the audio_outputs folder.
      parameters:
        - name: filename
          in: path
          required: true
          description: Name of the audio file to retrieve.
          schema:
            type: string
            example: "output.mp3"
      responses:
        '200':
          description: Audio file retrieved successfully.
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    GenerateAudioRequest:
      type: object
      properties:
        text:
          type: string
          description: The text input to be converted into speech.
          example: "Hello, this is a test audio."
      required:
        - text
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "No text provided."
        details:
          type: string
          example: "Failed to process the request."